fix : InternalTicketSettlementItemResponse DTO응답데이터 Long타입으로 수정

+++ commerce/src/main/java/com/devticket/commerce/ticket/presentation/dto/res/InternalTicketSettlementItemResponse.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 public record InternalTicketSettlementItemResponse(
     UUID eventId,
     UUID orderItemId,
-    int salesAmount,
-    int refundAmount
+    Long salesAmount,
+    Long refundAmount
 ) {
 
 }
